A few static_casts to compile with -Wc++11-narrowing

+++ include/QuarkLineBlock2.h
@@ -32,7 +32,7 @@ class QuarkLineBlock2 {
 
   std::vector<DilutedFactor> const &operator()(const int t,
                                                   const int b,
-                                                  const int op_id) const {
+                                                  const size_t op_id) const {
     auto const id =
         std::find(Ql_id.begin(), Ql_id.end(), std::pair<int, int>(t, b)) - Ql_id.begin();
     /*! @todo catch when t,b is an invalid index */
